% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{ricky_palette}
\alias{ricky_palette}
\title{Palette}
\usage{
ricky_palette(primary = "yellow", other = "blue", direction = 1)
}
\description{
The color palettes generated by ricky.
}
\details{
{
\item{primary}{The primary colour}
\item{other}{The other colour}
\item{direction}{= 1 is preface the palette colour, = -1 is reverse order the palette colour}
\item{n}{The number of colour select}
}
}
\examples{
# ricky_palette(direction = 1)(2)
}
