% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continous-scale.R
\name{scale_colour_ricky_seq_c}
\alias{scale_colour_ricky_seq_c}
\title{Continuous sequence Color Scales for ggplot2}
\usage{
scale_colour_ricky_seq_c(name, direction = 1, amount = 0.2, ...)
}
\description{
Continuous ggplot2 color scales using the color palettes generated by ricky
}
\details{
\describe{
\item{name}{The name of colour in palette}
\item{other}{The other colour}
\item{direction}{= 1 is preface the palette colour, = -1 is reverse order the palette colour}
\item{amount}{Colour shades}
}
}
\examples{
# plot with sequence scale gg + scale_colour_ricky_seq_c()
}
