% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-scale.R
\name{scale_fill_ricky_d}
\alias{scale_fill_ricky_d}
\title{Discrete fill scales for ggplot2}
\usage{
scale_fill_ricky_d(direction = 1, primary = "blue", other = "green", ...)
}
\description{
Discrete ggplot2 fill scales using the color palettes generated by ricky
}
\details{

}
\examples{
# plot with discrete scale gg + scale_fill_ricky_d()
}
