% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{ricky_palette}
\alias{ricky_palette}
\title{Palette}
\usage{
ricky_palette(primary = "yellow", other = "blue", direction = 1)
}
\description{
The color palettes generated by ricky.
}
\details{

}
\examples{
# ricky_palette(direction = 1)(2)
}
