% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-scale.R
\name{scale_fill_ricky_d}
\alias{scale_fill_ricky_d}
\title{Discrete fill scales for ggplot2}
\usage{
scale_fill_ricky_d(direction = 1, primary = "blue", other = "green", ...)
}
\description{
Discrete ggplot2 fill scales using the color palettes generated by ricky
}
\details{
\describe{
\item{direction}{= 1 is preface the palette colour, = -1 is reverse order the palette colour}
\item{primary}{The primary colour}
\item{other}{The other colour}
}
}
\examples{
# plot with discrete scale gg + scale_fill_ricky_d()
}
